pyproject.toml
[tool.poetry]
name = "repl_python3_R-Reaper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
colorama = "^0.4.4"
httpx = "^0.21.3"
pygame = "^2.1.2"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
